+++ nopol/src/main/resources/config.ini
@@ -13,6 +13,8 @@ maxLineInvocationPerTest=250
 timeoutMethodInvocation=2000
 onlyOneResult=true
 maxTimeBuildPatch=300
+# Maximal time allowed to each type ( precondition and condition ) before timeout
+maxTimeEachType=300
 
 [weight]
 methodCall=0.1
